Code refactor
Moved some files around to try and make a better system.

First of all, get rid of the "all classes are libraries" kind of thought. This is beyound useless and ridiculous. There are three actual libraries so far: random, system and text.

Second, get rid of "all classes in its corresponding directory". This was getting tiresome and by no means would be any better as the code grows.

Then, all that is left is reserve space for tests and remake CMakeLists.txt to reflect the changes. Everything works again, and maybe the code is a little more organized.

Just need to sort the dependency problem of the system library and that source dir in config.hpp.in. No idea how to do it now...